+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,14 +1,14 @@
 package kitchenpos.tablegroup.application;
 
-import kitchenpos.dao.TableGroupDao;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.dao.OrderTableDao;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTables;
-import kitchenpos.request.TableGroupCreateRequest;
-import kitchenpos.request.TableGroupUnitDto;
+import kitchenpos.tablegroup.dao.TableGroupDao;
 import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.request.TableGroupCreateRequest;
+import kitchenpos.tablegroup.request.TableGroupUnitDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
