+++ src/main/java/kitchenpos/tablegroup/dao/JdbcTemplateTableGroupDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.tablegroup.dao;
 
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
