+++ src/main/java/kitchenpos/tablegroup/request/TableGroupUnitDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.request;
+package kitchenpos.tablegroup.request;
 
 public class TableGroupUnitDto {
