+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,7 +1,6 @@
 package kitchenpos.application;
 
 import kitchenpos.common.domain.Price;
-import kitchenpos.product.dao.ProductDao;
 import kitchenpos.menu.dao.MenuDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
@@ -15,6 +14,7 @@ import kitchenpos.order.request.OrderCreateRequest;
 import kitchenpos.order.request.OrderLineItemDto;
 import kitchenpos.ordertable.dao.OrderTableDao;
 import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.product.dao.ProductDao;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
