+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -16,8 +16,8 @@ import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.request.OrderTableCreateRequest;
 import kitchenpos.ordertable.request.TableChangeEmptyRequest;
 import kitchenpos.ordertable.request.TableChangeNumberOfGuestsRequest;
-import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.request.TableGroupCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
